Add administrative order history

+++ jcart-administration-back/src/main/java/io/cjf/jcartadministrationback/po/Order.java
@@ -21,7 +21,7 @@ public class Order {
         return orderId;
     }
 
-    public void setOrderId(Long orderId) {
+    public void setOrderId(Integer orderId) {
         this.orderId = orderId;
     }
